requirements.txt
Babel==2.7.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-crispy-forms==1.8.1
django-phonenumber-field==3.0.1
idna==2.8
phonenumbers==8.11.0
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
